+++ packages/project/src/android/project.ts
@@ -149,9 +149,17 @@ export class AndroidProject extends PlatformProject {
    */
   async setPackageName(packageName: string) {
     const sourceDir = join(this.getAppRoot()!, 'src', 'main', 'java');
-    const oldPackageName = await this.manifest
+    let hadPackageAttr = false;
+    let oldPackageName = await this.manifest
       .getDocumentElement()
       ?.getAttribute('package');
+    
+    if (!oldPackageName) {
+      oldPackageName = await this.appBuildGradle?.getApplicationId();
+    } else {
+      hadPackageAttr = true;
+    }
+    
     const oldPackageParts = oldPackageName?.split('.') ?? [];
 
     Logger.v('android', 'setPackageName', 'setting Android package name to', packageName, 'from', oldPackageName);
@@ -167,12 +175,17 @@ export class AndroidProject extends PlatformProject {
       );
     }
 
-    this.manifest.getDocumentElement()?.setAttribute('package', packageName);
+    let activityName = '.MainActivity';
+    if (hadPackageAttr) {
+      this.manifest.getDocumentElement()?.setAttribute('package', packageName);
+      activityName = `${packageName}.MainActivity`;
+    }
+
     await this.appBuildGradle?.setApplicationId(packageName);
     await this.appBuildGradle?.setNamespace(packageName);
     Logger.v('android', 'setPackageName', `set manifest package attribute and applicationId to ${packageName}`);
     this.manifest.setAttrs('manifest/application/activity', {
-      'android:name': `${packageName}.MainActivity`,
+      'android:name': activityName
     });
     Logger.v('android', 'setPackageName', `set <activity android:name="${packageName}.MainActivity"`);
 
@@ -240,6 +253,14 @@ export class AndroidProject extends PlatformProject {
       return '';
     }
     return `${parts[parts.length - 1]}.java`;
+  }
+
+  async getMainActivityPath() {
+    const packageName = await this.appBuildGradle?.getApplicationId();
+    const activityName = this.getMainActivityFilename();
+    const packageParts = packageName?.split('.') ?? [];
+
+    return join('app', 'src', 'main', 'java', ...packageParts, activityName);
   }
 
   async getGradlePluginVersion() {