+++ packages/configure/CHANGELOG.md
@@ -1,5 +1,29 @@
 # @trapezedev/configure
 
+## 7.0.10
+
+### Patch Changes
+
+- 6dc56c2: Working on fixing android package renaming
+- Updated dependencies [6dc56c2]
+  - @trapezedev/project@7.0.10
+
+## 7.0.10-next.0
+
+### Patch Changes
+
+- Working on fixing android package renaming
+- Updated dependencies
+  - @trapezedev/project@7.0.10-next.0
+
+## 7.0.9
+
+### Patch Changes
+
+- Fixed package name detection and updated types
+- Updated dependencies
+  - @trapezedev/project@7.0.9
+
 ## 7.0.8
 
 ### Patch Changes